refactored table
