+++ frontend/src/components/assessment-creation/AssessmentTable.tsx
@@ -1,11 +1,15 @@
 import React from "react";
+import { Text } from "@chakra-ui/react";
 import { AssessmentType } from "../../types/AssessmentType";
-import { TableRow, Table } from "./Table";
 import EditAssessmentPopover from "./EditAssessmentPopover";
+import StatusTag from "./StatusTag";
+import { TableRow, Table } from "../common/Table";
 
 interface AssessmentTableProps {
   assessments: AssessmentType[];
 }
+
+type StatusProperty = "Draft" | "Published" | "Archived" | "Deleted";
 
 const AssessmentTable = ({
   assessments,
@@ -13,8 +17,10 @@ const AssessmentTable = ({
   const headers = ["Status", "Name", "Grade", "Type", "Country", "Region"];
   const rows: TableRow[] = assessments.map((assessment) => ({
     values: [
-      assessment.status,
-      assessment.name,
+      <StatusTag key={1} status={assessment.status as StatusProperty} />,
+      <Text key={2} fontWeight="bold">
+        {assessment.name}
+      </Text>,
       assessment.grade,
       assessment.type,
       assessment.country,
