+++ frontend/src/components/user-management/AdminUserTable.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { Text } from "@chakra-ui/react";
 import { AdminUser } from "../../types/UserTypes";
 import { TableRow, Table } from "../common/Table";
 import RemoveUserPopover from "./RemoveUserPopover";
@@ -12,7 +13,12 @@ const AdminUserTable = ({
 }: AdminUserTableProps): React.ReactElement => {
   const headers = ["Name", "Email"];
   const rows: TableRow[] = adminUsers.map((user) => ({
-    values: [`${user.firstName} ${user.lastName}`, user.email],
+    values: [
+      <Text key={0} fontWeight="bold">
+        {user.firstName} {user.lastName}
+      </Text>,
+      user.email,
+    ],
     menu: (
       <RemoveUserPopover
         name={`${user.firstName} ${user.lastName}`}
